move cq infra to separate class
